Replace `IndexUid::new` with `IndexUid::from` in metastore shard API impl
`IndexUid::new` sets the incarnation ID to `Ulid::Nil()`.
