+++ quickwit/quickwit-proto/src/metastore/mod.rs
@@ -217,13 +217,6 @@ impl SourceType {
 }
 
 impl IndexMetadataRequest {
-    pub fn for_index_uid(index_uid: IndexUid) -> Self {
-        Self {
-            index_uid: Some(index_uid.into()),
-            index_id: None,
-        }
-    }
-
     pub fn for_index_id(index_id: IndexId) -> Self {
         Self {
             index_uid: None,
@@ -231,9 +224,16 @@ impl IndexMetadataRequest {
         }
     }
 
+    pub fn for_index_uid(index_uid: IndexUid) -> Self {
+        Self {
+            index_uid: Some(index_uid.into()),
+            index_id: None,
+        }
+    }
+
     /// Returns the index id either from the `index_id` or the `index_uid`.
     /// If none of them is set, an error is returned.
-    pub fn get_index_id(&self) -> MetastoreResult<String> {
+    pub fn get_index_id(&self) -> MetastoreResult<IndexId> {
         if let Some(index_id) = &self.index_id {
             Ok(index_id.to_string())
         } else if let Some(index_uid) = &self.index_uid {
@@ -249,9 +249,9 @@ impl IndexMetadataRequest {
 }
 
 impl MarkSplitsForDeletionRequest {
-    pub fn new(index_uid: String, split_ids: Vec<String>) -> Self {
+    pub fn new(index_uid: IndexUid, split_ids: Vec<String>) -> Self {
         Self {
-            index_uid,
+            index_uid: index_uid.into(),
             split_ids,
         }
     }
@@ -266,9 +266,9 @@ impl LastDeleteOpstampResponse {
 }
 
 impl ListDeleteTasksRequest {
-    pub fn new(index_uid: String, opstamp_start: u64) -> Self {
+    pub fn new(index_uid: IndexUid, opstamp_start: u64) -> Self {
         Self {
-            index_uid,
+            index_uid: index_uid.into(),
             opstamp_start,
         }
     }
